arduino/requirements.txt
Flask
flask-socketio
pyserial
flask-cors
